use simple trigger for OneOffJobBootstrap
