+++ elastic-job-lite-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/api/bootstrap/JobBootstrap.java
@@ -23,27 +23,48 @@ import org.apache.shardingsphere.elasticjob.lite.api.listener.AbstractDistribute
 import org.apache.shardingsphere.elasticjob.lite.api.listener.ElasticJobListener;
 import org.apache.shardingsphere.elasticjob.lite.api.script.ScriptJob;
 import org.apache.shardingsphere.elasticjob.lite.config.JobConfiguration;
+import org.apache.shardingsphere.elasticjob.lite.exception.JobSystemException;
 import org.apache.shardingsphere.elasticjob.lite.handler.sharding.JobInstance;
 import org.apache.shardingsphere.elasticjob.lite.internal.guarantee.GuaranteeService;
 import org.apache.shardingsphere.elasticjob.lite.internal.schedule.JobRegistry;
+import org.apache.shardingsphere.elasticjob.lite.internal.schedule.JobScheduleController;
+import org.apache.shardingsphere.elasticjob.lite.internal.schedule.JobShutdownHookPlugin;
+import org.apache.shardingsphere.elasticjob.lite.internal.schedule.LiteJob;
 import org.apache.shardingsphere.elasticjob.lite.internal.schedule.SchedulerFacade;
 import org.apache.shardingsphere.elasticjob.lite.internal.setup.SetUpFacade;
 import org.apache.shardingsphere.elasticjob.lite.reg.base.CoordinatorRegistryCenter;
 import org.apache.shardingsphere.elasticjob.lite.tracing.api.TracingConfiguration;
+import org.quartz.JobBuilder;
+import org.quartz.JobDetail;
+import org.quartz.Scheduler;
+import org.quartz.SchedulerException;
+import org.quartz.impl.StdSchedulerFactory;
+import org.quartz.simpl.SimpleThreadPool;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Properties;
 
 /**
  * Schedule job bootstrap.
  */
-@Getter
 public abstract class JobBootstrap {
+    
+    private static final String REG_CENTER_DATA_MAP_KEY = "regCenter";
+    
+    private static final String ELASTIC_JOB_DATA_MAP_KEY = "elasticJob";
+    
+    private static final String JOB_CONFIG_DATA_MAP_KEY = "jobConfig";
+    
+    private static final String JOB_LISTENERS_DATA_MAP_KEY = "elasticJobListeners";
+    
+    private static final String TRACING_CONFIG_DATA_MAP_KEY = "tracingConfig";
     
     private final CoordinatorRegistryCenter regCenter;
     
     private final ElasticJob elasticJob;
     
+    @Getter
     private final JobConfiguration jobConfig;
     
     private final List<ElasticJobListener> elasticJobListeners;
@@ -79,11 +100,54 @@ public abstract class JobBootstrap {
         }
     }
     
-    protected void registerStartUpInfo() {
+    protected final JobScheduleController createJobScheduleController() {
+        JobScheduleController result = new JobScheduleController(createScheduler(), createJobDetail(), getJobConfig().getJobName());
+        JobRegistry.getInstance().registerJob(getJobConfig().getJobName(), result);
+        registerStartUpInfo();
+        return result;
+    }
+    
+    private void registerStartUpInfo() {
         JobRegistry.getInstance().registerRegistryCenter(jobConfig.getJobName(), regCenter);
         JobRegistry.getInstance().addJobInstance(jobConfig.getJobName(), new JobInstance());
         JobRegistry.getInstance().setCurrentShardingTotalCount(jobConfig.getJobName(), jobConfig.getShardingTotalCount());
         setUpFacade.registerStartUpInfo(!jobConfig.isDisabled());
+    }
+    
+    private Scheduler createScheduler() {
+        Scheduler result;
+        try {
+            StdSchedulerFactory factory = new StdSchedulerFactory();
+            factory.initialize(getQuartzProps());
+            result = factory.getScheduler();
+            result.getListenerManager().addTriggerListener(schedulerFacade.newJobTriggerListener());
+        } catch (final SchedulerException ex) {
+            throw new JobSystemException(ex);
+        }
+        return result;
+    }
+    
+    private Properties getQuartzProps() {
+        Properties result = new Properties();
+        result.put("org.quartz.threadPool.class", SimpleThreadPool.class.getName());
+        result.put("org.quartz.threadPool.threadCount", "1");
+        result.put("org.quartz.scheduler.instanceName", getJobConfig().getJobName());
+        result.put("org.quartz.jobStore.misfireThreshold", "1");
+        result.put("org.quartz.plugin.shutdownhook.class", JobShutdownHookPlugin.class.getName());
+        result.put("org.quartz.plugin.shutdownhook.cleanShutdown", Boolean.TRUE.toString());
+        return result;
+    }
+    
+    private JobDetail createJobDetail() {
+        JobDetail result = JobBuilder.newJob(LiteJob.class).withIdentity(getJobConfig().getJobName()).build();
+        result.getJobDataMap().put(REG_CENTER_DATA_MAP_KEY, regCenter);
+        result.getJobDataMap().put(JOB_CONFIG_DATA_MAP_KEY, getJobConfig());
+        result.getJobDataMap().put(JOB_LISTENERS_DATA_MAP_KEY, elasticJobListeners);
+        result.getJobDataMap().put(TRACING_CONFIG_DATA_MAP_KEY, tracingConfig);
+        if (null != elasticJob && !elasticJob.getClass().getName().equals(ScriptJob.class.getName())) {
+            result.getJobDataMap().put(ELASTIC_JOB_DATA_MAP_KEY, elasticJob);
+        }
+        return result;
     }
     
    /**
